2020-04-23 Bug fix in DialogSample; Added InputValidation to CascadingDefaults
