+++ BlazorMdc.Demo.CommonUI/Pages/CascadingDefaultsSample/CascadingDefaultsSample.razor
@@ -58,7 +58,7 @@
         SelectInputStyle = MdcSelectInputStyle.Outlined,
         TextInputStyle = MdcTextInputStyle.Outlined,
         ButtonStyle = MdcButtonStyle.Outlined,
-        DateSelectionCriteria = MdcDateSelectionCriteria.WeekdaysOnly
+        DateSelectionCriteria = PMdcDateSelectionCriteria.WeekdaysOnly
     };
     string selectVariant1;
     string textFieldVariant1;
@@ -71,7 +71,7 @@
         SelectInputStyle = MdcSelectInputStyle.Filled,
         TextInputStyle = MdcTextInputStyle.FullWidth,
         ButtonStyle = MdcButtonStyle.ContainedUnelevated,
-        DateSelectionCriteria = MdcDateSelectionCriteria.WeekendsOnly
+        DateSelectionCriteria = PMdcDateSelectionCriteria.WeekendsOnly
     };
     string selectVariant2;
     string textFieldVariant2;
@@ -84,7 +84,7 @@
         SelectInputStyle = MdcSelectInputStyle.Outlined,
         TextInputStyle = MdcTextInputStyle.Filled,
         ButtonStyle = MdcButtonStyle.ContainedRaised,
-        DateSelectionCriteria = MdcDateSelectionCriteria.AllowAll
+        DateSelectionCriteria = PMdcDateSelectionCriteria.AllowAll
     };
     string selectVariant3;
     string textFieldVariant3;
