+++ BlazorMdc.Demo.CommonUI/Pages/DatePickerSample/DatePickerSample.razor
@@ -11,7 +11,7 @@
 <PMdcDatePicker Label="Filled" @bind-Value="@filled" SelectInputStyle="MdcSelectInputStyle.Filled" MinDate="min" MaxDate="max"></PMdcDatePicker>
 <p style="margin-bottom: 4em;">Selected: '<strong>@filled.ToString()</strong>'</p>
 
-<PMdcDatePicker Label="Weekdays Outlined" @bind-Value="@outlined" SelectInputStyle="MdcSelectInputStyle.Outlined" DateSelectionCriteria="MdcDateSelectionCriteria.WeekdaysOnly" MinDate="min" MaxDate="max"></PMdcDatePicker>
+<PMdcDatePicker Label="Weekdays Outlined" @bind-Value="@outlined" SelectInputStyle="MdcSelectInputStyle.Outlined" DateSelectionCriteria="PMdcDateSelectionCriteria.WeekdaysOnly" MinDate="min" MaxDate="max"></PMdcDatePicker>
 <p style="margin-bottom: 20em;">Selected: '<strong>@outlined.ToString()</strong>'</p>
 
 <p>Inline date picker purely for test purposes - in production please use <code>PMdcDatePicker</code></p>
