+++ BlazorMdc/Base/MdcCascadingDefaults.cs
@@ -1,63 +1,9 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
+﻿//
+//  2020-04-23  Mark Stega
+//              Removed all enumerations and placed them in MdcEnumerations
+//
 namespace BlazorMdc
 {
-    /// <summary>
-    /// The MDC Top App Bar type: all items are mutually exclusive
-    /// </summary>
-    public enum MdcTopAppBarType { Standard, Fixed, Dense, Prominent, Short, ShortCollapsed }
-
-
-    /// <summary>
-    /// The style of MDC input: note that FullWidth is mutually exclusive with Outlined
-    /// </summary>
-    public enum MdcTextAlignStyle { Default, Left, Center, Right }
-
-
-    /// <summary>
-    /// The style of MDC input: note that FullWidth is mutually exclusive with Outlined
-    /// </summary>
-    public enum MdcTextInputStyle { Filled, Outlined, FullWidth }
-
-
-    /// <summary>
-    /// The style of MDC select
-    /// </summary>
-    public enum MdcSelectInputStyle { Filled, Outlined }
-
-
-    /// <summary>
-    /// The style of MDC input: note that FullWidth is mutually exclusive with Outlined
-    /// </summary>
-    public enum MdcNumericInputMagnitude { Normal = 0, Percent = 2, BasisPoints = 4 }
-
-
-    /// <summary>
-    /// MDC button styling
-    /// </summary>
-    public enum MdcButtonStyle { ContainedRaised, ContainedUnelevated, Outlined, Text }
-
-
-    /// <summary>
-    /// MDC list styling
-    /// </summary>
-    public enum MdcListStyle { None, Outlined }
-
-
-    /// <summary>
-    /// MDC card styling
-    /// </summary>
-    public enum MdcCardStyle { Default, Outlined }
-
-
-    /// <summary>
-    /// Date Picker selection criteria
-    /// </summary>
-    public enum MdcDateSelectionCriteria { AllowAll, WeekendsOnly, WeekdaysOnly }
-
-
     public class MdcCascadingDefaults
     {
         public MdcButtonStyle ButtonStyle { get; set; } = MdcButtonStyle.Text;
@@ -70,7 +16,8 @@ namespace BlazorMdc
         public MdcTextInputStyle TextInputStyle { get; set; } = MdcTextInputStyle.Filled;
         public MdcSelectInputStyle SelectInputStyle { get; set; } = MdcSelectInputStyle.Filled;
 
-        public MdcDateSelectionCriteria DateSelectionCriteria { get; set; } = MdcDateSelectionCriteria.AllowAll;
+        public PMdcDateSelectionCriteria DateSelectionCriteria { get; set; } = PMdcDateSelectionCriteria.AllowAll;
+        public PMdcRadioButtonGroupItemValidation RadioButtonGroupItemValidation { get; set; } = PMdcRadioButtonGroupItemValidation.Exception;
 
         public string TextFieldCssClass { get; set; } = "";
 
@@ -85,6 +32,7 @@ namespace BlazorMdc
         public MdcTextInputStyle AppliedStyle(MdcTextInputStyle? style = null) => (style is null) ? TextInputStyle : (MdcTextInputStyle)style;
         public MdcSelectInputStyle AppliedStyle(MdcSelectInputStyle? style = null) => (style is null) ? SelectInputStyle : (MdcSelectInputStyle)style;
 
-        public MdcDateSelectionCriteria AppliedDateSelectionCriteria(MdcDateSelectionCriteria? criteria = null) => (criteria is null) ? DateSelectionCriteria : (MdcDateSelectionCriteria)criteria;
+        public PMdcDateSelectionCriteria AppliedDateSelectionCriteria(PMdcDateSelectionCriteria? criteria = null) => (criteria is null) ? DateSelectionCriteria : (PMdcDateSelectionCriteria)criteria;
+        public PMdcRadioButtonGroupItemValidation AppliedRadioButtonGroupItemValidation(PMdcRadioButtonGroupItemValidation? criteria = null) => (criteria is null) ? RadioButtonGroupItemValidation : (PMdcRadioButtonGroupItemValidation)criteria;
     }
 }
