+++ BlazorMdc/Plus/PMdcDatePicker/InternalDatePickerPanel.razor
@@ -81,7 +81,7 @@
     //  2020-04-02  Mark Stega
     //              Changed to use class & style mappers
     //
-    [Parameter] public MdcDateSelectionCriteria? DateSelectionCriteria { get; set; }
+    [Parameter] public PMdcDateSelectionCriteria? DateSelectionCriteria { get; set; }
     [Parameter] public DateTime MinDate { get; set; }
     [Parameter] public DateTime MaxDate { get; set; }
