+++ BlazorMdc/Plus/PMdcDatePicker/PMdcDatePicker.razor
@@ -47,7 +47,7 @@
     //
     [Parameter] public MdcSelectInputStyle? SelectInputStyle { get; set; }
     [Parameter] public string Label { get; set; }
-    [Parameter] public MdcDateSelectionCriteria? DateSelectionCriteria { get; set; }
+    [Parameter] public PMdcDateSelectionCriteria? DateSelectionCriteria { get; set; }
     [Parameter] public DateTime MinDate { get; set; }
     [Parameter] public DateTime MaxDate { get; set; }
