+++ BlazorMdc/Plus/PMdcRadioButtonGroup/PMdcRadioButtonGroup.razor
@@ -16,7 +16,7 @@
             <div class="bmdc-mdc-radio-break"></div>
         }
 
-        <MdcRadioButton @key="@item"
+        <MdcRadioButton @key="@index"
                         Value="@Value"
                         ValueChanged="@ValueChanged"
                         ValueExpression="@ValueExpression"
@@ -24,8 +24,8 @@
                         Disabled="@item.Disabled"
                         EnableTouchWrapper="@EnableTouchWrapper"
                         Label="@item.Label"
-                        CheckedValue="@item.DataValue"
-                        @onclick="@(_ => OnItemClickAsync(item.DataValue))" />
+                        CheckedValue="@item.CheckedValue"
+                        @onclick="@(_ => OnItemClickAsync(item.CheckedValue))" />
     }
 
 </div>
@@ -37,7 +37,7 @@
     //
     public class Item<T>
     {
-        public T DataValue { get; set; }
+        public T CheckedValue { get; set; }
         public string Label { get; set; }
         public string ButtonContainerClass { get; set; }
         public bool Disabled { get; set; } = false;
@@ -45,7 +45,7 @@
 
     [Parameter] public IEnumerable<Item<TItem>> Items { get; set; }
     [Parameter] public bool Vertical { get; set; } = false;
-    [Parameter] public RadioButtonGroup.MdcRadioButtonGroupItemValidation ItemValidation { get; set; } = RadioButtonGroup.MdcRadioButtonGroupItemValidation.Exception;
+    [Parameter] public PMdcRadioButtonGroupItemValidation? ItemValidation { get; set; }
     [Parameter] public bool EnableTouchWrapper { get; set; } = true;
 
 
@@ -58,26 +58,28 @@
             throw new ArgumentException("PMdcDialogButtonGroup requires a non-empty Items parameter.");
         }
 
-        if ((ItemValidation == RadioButtonGroup.MdcRadioButtonGroupItemValidation.DefaultToFirst) && (Value is null || string.IsNullOrEmpty(Value.ToString())))
+        PMdcRadioButtonGroupItemValidation appliedItemValidation = CascadingDefaults.AppliedRadioButtonGroupItemValidation(ItemValidation);
+
+        if ((appliedItemValidation == PMdcRadioButtonGroupItemValidation.DefaultToFirst) && (Value is null || string.IsNullOrEmpty(Value.ToString())))
         {
-            await OnItemClickAsync(Items.FirstOrDefault().DataValue);
+            await OnItemClickAsync(Items.FirstOrDefault().CheckedValue);
         }
 
-        if (Items.Where(i => object.Equals(i.DataValue, Value)).Count() == 0)
+        if (Items.Where(i => object.Equals(i.CheckedValue, Value)).Count() != 1)
         {
-            switch (ItemValidation)
+            switch (appliedItemValidation)
             {
-                case RadioButtonGroup.MdcRadioButtonGroupItemValidation.DefaultToFirst:
-                    await OnItemClickAsync(Items.FirstOrDefault().DataValue);
+                case PMdcRadioButtonGroupItemValidation.DefaultToFirst:
+                    await OnItemClickAsync(Items.FirstOrDefault().CheckedValue);
                     break;
 
-                case RadioButtonGroup.MdcRadioButtonGroupItemValidation.Exception:
+                case PMdcRadioButtonGroupItemValidation.Exception:
                     string itemList = "{ ";
                     string prepend = "";
 
                     foreach (var item in Items)
                     {
-                        itemList += $"{prepend} '{item.DataValue}'";
+                        itemList += $"{prepend} '{item.CheckedValue}'";
                         prepend = ",";
                     }
 
@@ -85,15 +87,15 @@
 
                     throw new ArgumentException($"PMdcRadioButtonGroup cannot select item with data value of '{Value?.ToString()}' from {itemList}");
 
-                case RadioButtonGroup.MdcRadioButtonGroupItemValidation.NoSelection:
+                case PMdcRadioButtonGroupItemValidation.NoSelection:
                     break;
             }
         }
     }
 
-    private async Task OnItemClickAsync(TItem dataValue)
+    private async Task OnItemClickAsync(TItem CheckedValue)
     {
-        Value = dataValue;
+        Value = CheckedValue;
         await ValueChanged.InvokeAsync(Value);
         StateHasChanged();
     }
