--- BlazorMdc/Plus/PMdcRadioButtonGroup/RadioButtonGroup.cs
@@ -1,7 +1,0 @@
-﻿namespace BlazorMdc
-{
-    public class RadioButtonGroup
-    {
-        public enum MdcRadioButtonGroupItemValidation { DefaultToFirst, Exception, NoSelection }
-    }
-}
